moved bulk upload to proper route
